Code refactor (#6)
* path to PR_TEMPLATE fix

* path to Issue template fix

* code refactor

* relative module paths instead of global paths
