+++ src/features/common/index.ts
@@ -1,0 +1,3 @@
+import { compareTemplates } from './compareTemplates'
+
+export { compareTemplates }
