+++ src/features/issueTemplate/index.ts
@@ -1,0 +1,4 @@
+import { findExistingDefaultIssueTemplate } from './findExistingIssueTemplate'
+import { checkDefaultIssueTemplate } from './checkDefaultIssueTemplate'
+
+export { findExistingDefaultIssueTemplate, checkDefaultIssueTemplate }
