+++ src/features/prTemplate/index.ts
@@ -1,0 +1,4 @@
+import { findExistingPRTemplate } from './findExistingPrTemplate'
+import { checkPRTemplate } from './checkPrTemplate'
+
+export { findExistingPRTemplate, checkPRTemplate }
